+++ test/mapbox_search_test.dart
@@ -1,8 +1,9 @@
 import 'package:mapbox_search/mapbox_search.dart';
 import 'package:test/test.dart';
+import 'credentials.dart' as c;
 
 void main() async {
-  final MAPBOX_KEY = '';
+  final MAPBOX_KEY = c.MAPBOX_KEY;
 
   test("Has api key", () {
     expect(MAPBOX_KEY, isNotNull);
@@ -10,15 +11,16 @@ void main() async {
   });
 
   test("Places search test", () async {
-    var search = PlacesSearch(
+    PlacesSearch search = PlacesSearch(
       apiKey: MAPBOX_KEY,
       country: "BR",
       limit: 5,
+      types: [PlaceType.address, PlaceType.place],
     );
 
     var searchPlace = search.getPlaces(
-      "patio",
-      location: Location(
+      "central park",
+      proximity: Location(
         lat: -19.984634,
         lng: -43.9502958,
       ),